Use native nextjs proxy for graphql
