+++ .env
@@ -2,4 +2,4 @@
 # Please, use a `.env.local` file for you local config
 # See also: https://nextjs.org/docs/app/building-your-application/configuring/environment-variables
 
-GUACGQL_SERVER_URL=http://localhost:8080
+NEXT_PUBLIC_GUACGQL_SERVER_URL=http://localhost:8080
